gadgetry_backend/requirements.txt
asgiref==3.6.0
dj-database-url==2.0.0
Django==4.1.7
django-cors-headers==3.14.0
django-cron==0.6.0
django-filter==22.1
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
whitenoise==6.4.0
